delete unused ProfilePage
